Create project9
Purpose of the program is ask user what to draw.Based 
on selection, the screen will display an option with 
numbers of pixel the screen should fill up with
their first choosen option. The image will be drawn 
from the center of the screen. 

Ex: Op 1) Draw circle. Op 2) Radius 10 pixels from center.